# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  model.Cargo:
    properties:
      carga_horaria_diaria_minutos:
        type: integer
      empresa_id:
        type: integer
      entrada_esperada_minutos:
        type: integer
      id:
        type: integer
      minutos_almoco_esperado:
        type: integer
      nome:
        type: string
      permissoes:
        items:
          $ref: '#/definitions/model.Permissao'
        type: array
      saida_esperada_minutos:
        type: integer
      salario_maximo:
        type: number
      salario_minimo:
        type: number
    type: object
  model.Contrato:
    properties:
      cargo:
        $ref: '#/definitions/model.Cargo'
      cargo_id:
        type: integer
      data_admissao:
        type: string
      data_demissao:
        type: string
      empresa:
        $ref: '#/definitions/model.Empresa'
      empresa_id:
        type: integer
      id:
        type: integer
      localidade:
        $ref: '#/definitions/model.Localidade'
      localidade_id:
        type: integer
      salario:
        type: number
      saldo_banco_horas_minutos:
        type: integer
      usuario_id:
        type: integer
    type: object
  model.Empresa:
    properties:
      cnpj:
        type: string
      id:
        type: integer
      localidades:
        items:
          $ref: '#/definitions/model.Localidade'
        type: array
      nome_fantasia:
        type: string
      razao_social:
        type: string
    type: object
...
  model.Localidade:
    properties:
      bairro:
        type: string
      cep:
        type: string
      cidade:
        type: string
      empresa_id:
        type: integer
      estado:
        type: string
      id:
        type: integer
      latitude:
        type: number
      logradouro:
        type: string
      longitude:
        type: number
      nome:
        description: 'Ex: "Matriz São Paulo", "Filial Rio"'
        type: string
      numero:
        type: string
      raio_geofence_metros:
        type: number
    type: object
  model.Permissao:
    properties:
      descricao:
        type: string
      id:
        type: integer
      nome:
        type: string
    type: object
...
  model.Usuario:
    properties:
      contrato:
        $ref: '#/definitions/model.Contrato'
      cpf:
        type: string
      data_atualizacao:
        type: string
      data_criacao:
        type: string
      email:
        type: string
      id:
        type: integer
      nome:
        type: string
    type: object
...
  usuario.CriarUsuarioRequest:
    properties:
      cargo_id:
        type: integer
      cpf:
        type: string
      data_admissao:
        type: string
      email:
        type: string
      empresa_id:
        description: Dados do Contrato
        type: integer
      localidade_id:
        type: integer
      nome:
        description: Dados do Usuário (Pessoa)
        type: string
      salario:
        type: number
      senha:
        minLength: 6
        type: string
    required:
    - cargo_id
    - cpf
    - data_admissao
    - email
    - empresa_id
    - localidade_id
    - nome
    - salario
    - senha
    type: object
  usuario.UpdateUsuarioRequest:
    properties:
      email:
        example: joao.dasilva@empresa.com
        type: string
      nome:
        example: João da Silva
        type: string
      senha:
        example: novaSenha456
        type: string
    type: object
host: localhost:8083
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: API de alta performance para um sistema de Ponto Eletrônico.
  license:
    name: MIT
    url: https://github.com/Loviiin/ponto-api-go/blob/main/LICENSE
  termsOfService: http://swagger.io/terms/
  title: Ponto API em Go
  version: "1.0"
paths:
...
  /usuarios:
    get:
      description: Retorna uma lista de todos os usuários pertencentes à empresa do
        requisitante.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.Usuario'
            type: array
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Lista todos os usuários da empresa
      tags:
      - Usuários
    post:
      consumes:
      - application/json
      description: Cria um novo usuário (funcionário) e seu contrato de trabalho no
        sistema.
      parameters:
      - description: Dados do Novo Usuário e Contrato
        in: body
        name: usuario
        required: true
        schema:
          $ref: '#/definitions/usuario.CriarUsuarioRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/model.Usuario'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Cria um novo usuário
      tags:
      - Usuários
  /usuarios/{id}:
    delete:
      description: Deleta um usuário. Requer permissão de 'DELETAR_USUARIO' ou 'DELETAR_PROPRIA_CONTA'.
      parameters:
      - description: ID do Usuário a ser deletado
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Deleta um usuário
      tags:
      - Usuários
    get:
      description: Retorna os dados de um usuário específico da mesma empresa.
      parameters:
      - description: ID do Usuário
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Usuario'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Busca um usuário por ID
      tags:
      - Usuários
    put:
      consumes:
      - application/json
      description: Atualiza os dados de um usuário. O corpo do pedido pode conter
        qualquer um dos campos definidos no modelo.
      parameters:
      - description: ID do Usuário a ser atualizado
        in: path
        name: id
        required: true
        type: integer
      - description: Dados para atualização
        in: body
        name: dados
        required: true
        schema:
          $ref: '#/definitions/usuario.UpdateUsuarioRequest'
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Atualiza um usuário
      tags:
      - Usuários
  /usuarios/me:
    get:
      description: Retorna as informações detalhadas do usuário que está a fazer o
        pedido.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Usuario'
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Obtém os dados do usuário logado
      tags:
      - Usuários
securityDefinitions:
  BearerAuth:
    description: '"Digite ''Bearer'' seguido de um espaço e o seu token."'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"